+++ cgv_clone/src/main/java/com/ceos22/cgv_clone/common/config/JpaConfig.java
@@ -1,0 +1,9 @@
+package com.ceos22.cgv_clone.common.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+}
